Infraestructura/IAWS/requirements.txt
pymongo==3.10.1
tornado==4.3
python-dateutil==2.5.3
uuid==1.30
pyjwt==2.1.0